refactor: Remove last overloading of 'server' module name (#653)
* refactor: Remove last duplicate 'server'

* refactor: move names to the data_types crate
